+++ src/file.ts
@@ -24,9 +24,7 @@ import {
 import {promisifyAll} from '@google-cloud/promisify';
 
 import compressible = require('compressible');
-import getStream = require('get-stream');
 import * as crypto from 'crypto';
-import * as dateFormat from 'date-and-time';
 import * as extend from 'extend';
 import * as fs from 'fs';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
@@ -64,7 +62,12 @@ import {
 } from './nodejs-common/util';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const duplexify: DuplexifyConstructor = require('duplexify');
-import {normalize, objectKeyToLowercase, unicodeJSONStringify} from './util';
+import {
+  normalize,
+  objectKeyToLowercase,
+  unicodeJSONStringify,
+  formatAsUTCISO,
+} from './util';
 import retry = require('async-retry');
 
 export type GetExpirationDateResponse = [Date];
@@ -1375,8 +1378,8 @@ class File extends ServiceObject<File> {
       ) => {
         if (err) {
           // Get error message from the body.
-          getStream(rawResponseStream).then(body => {
-            err.message = body;
+          this.getBufferFromReadable(rawResponseStream).then(body => {
+            err.message = body.toString('utf8');
             throughStream.destroy(err);
           });
 
@@ -2033,10 +2036,9 @@ class File extends ServiceObject<File> {
         fileStream.pipe(writable).on('error', callback).on('finish', callback);
       });
     } else {
-      getStream
-        .buffer(fileStream)
+      this.getBufferFromReadable(fileStream)
         .then(contents => callback?.(null, contents))
-        .catch(callback as (error: RequestError) => void);
+        .catch(callback as (err: RequestError) => void);
     }
   }
 
@@ -2506,8 +2508,8 @@ class File extends ServiceObject<File> {
     let fields = Object.assign({}, options.fields);
 
     const now = new Date();
-    const nowISO = dateFormat.format(now, 'YYYYMMDD[T]HHmmss[Z]', true);
-    const todayISO = dateFormat.format(now, 'YYYYMMDD', true);
+    const nowISO = formatAsUTCISO(now, true);
+    const todayISO = formatAsUTCISO(now);
 
     const sign = async () => {
       const {client_email} = await this.storage.authClient.getCredentials();
@@ -2532,11 +2534,7 @@ class File extends ServiceObject<File> {
 
       delete fields.bucket;
 
-      const expiration = dateFormat.format(
-        expires,
-        'YYYY-MM-DD[T]HH:mm:ss[Z]',
-        true
-      );
+      const expiration = formatAsUTCISO(expires, true, '-', ':');
 
       const policy = {
         conditions,
@@ -3845,6 +3843,15 @@ class File extends ServiceObject<File> {
       this.storage.retryOptions.autoRetry = false;
     }
   }
+
+  private async getBufferFromReadable(readable: Readable): Promise<Buffer> {
+    const buf = [];
+    for await (const chunk of readable) {
+      buf.push(chunk);
+    }
+
+    return Buffer.concat(buf);
+  }
 }
 
 /*! Developer Documentation
@@ -3859,6 +3866,7 @@ promisifyAll(File, {
     'save',
     'setEncryptionKey',
     'shouldRetryBasedOnPreconditionAndIdempotencyStrat',
+    'getBufferFromReadable',
   ],
 });