+++ test/resumable-upload.ts
@@ -50,6 +50,8 @@ const RESUMABLE_INCOMPLETE_STATUS_CODE = 308;
 /** 256 KiB */
 const CHUNK_SIZE_MULTIPLE = 2 ** 18;
 const queryPath = '/?userProject=user-project-id';
+const X_GOOG_API_HEADER_REGEX =
+  /^gl-node\/(?<nodeVersion>[^W]+) gccl\/(?<gccl>[^W]+) gccl-invocation-id\/(?<gcclInvocationId>[^W]+)$/;
 
 function mockAuthorizeRequest(
   code = 200,
@@ -88,7 +90,7 @@ describe('resumable-upload', () => {
     retryableErrorFn: RETRYABLE_ERR_FN_DEFAULT,
   };
   let REQ_OPTS: GaxiosOptions;
-  const keyFile = path.join(__dirname, 'fixtures', 'keys.json');
+  const keyFile = path.join(__dirname, '../../test/fixtures/keys.json');
 
   before(() => {
     mockery.registerMock('abort-controller', {default: AbortController});
@@ -848,6 +850,15 @@ describe('resumable-upload', () => {
           done();
         });
       });
+
+      it('currentInvocationId.uri should remain the same on error', done => {
+        const beforeCallInvocationId = up.currentInvocationId.uri;
+        up.createURI((err: Error) => {
+          assert(err);
+          assert.equal(beforeCallInvocationId, up.currentInvocationId.uri);
+          done();
+        });
+      });
     });
 
     describe('success', () => {
@@ -881,6 +892,14 @@ describe('resumable-upload', () => {
         up.createURI((err: Error, uri: string) => {
           assert.ifError(err);
           assert.strictEqual(uri, URI);
+          done();
+        });
+      });
+
+      it('currentInvocationId.uri should be different after success', done => {
+        const beforeCallInvocationId = up.currentInvocationId.uri;
+        up.createURI(() => {
+          assert.notEqual(beforeCallInvocationId, up.currentInvocationId.uri);
           done();
         });
       });
@@ -1050,9 +1069,14 @@ describe('resumable-upload', () => {
 
           await up.startUploading();
 
-          assert.deepEqual(reqOpts.headers, {
-            'Content-Range': `bytes ${OFFSET}-*/${CONTENT_LENGTH}`,
-          });
+          assert(reqOpts.headers);
+          assert.equal(
+            reqOpts.headers['Content-Range'],
+            `bytes ${OFFSET}-*/${CONTENT_LENGTH}`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+          );
 
           const data = await getAllDataFromRequest();
 
@@ -1064,9 +1088,11 @@ describe('resumable-upload', () => {
 
           await up.startUploading();
 
-          assert.deepEqual(reqOpts.headers, {
-            'Content-Range': 'bytes 0-*/*',
-          });
+          assert(reqOpts.headers);
+          assert.equal(reqOpts.headers['Content-Range'], 'bytes 0-*/*');
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+          );
 
           const data = await getAllDataFromRequest();
 
@@ -1091,10 +1117,15 @@ describe('resumable-upload', () => {
           await up.startUploading();
 
           const endByte = OFFSET + CHUNK_SIZE - 1;
-          assert.deepEqual(reqOpts.headers, {
-            'Content-Length': CHUNK_SIZE,
-            'Content-Range': `bytes ${OFFSET}-${endByte}/${CONTENT_LENGTH}`,
-          });
+          assert(reqOpts.headers);
+          assert.equal(reqOpts.headers['Content-Length'], CHUNK_SIZE);
+          assert.equal(
+            reqOpts.headers['Content-Range'],
+            `bytes ${OFFSET}-${endByte}/${CONTENT_LENGTH}`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+          );
 
           const data = await getAllDataFromRequest();
 
@@ -1110,10 +1141,15 @@ describe('resumable-upload', () => {
           await up.startUploading();
 
           const endByte = OFFSET + CHUNK_SIZE - 1;
-          assert.deepEqual(reqOpts.headers, {
-            'Content-Length': CHUNK_SIZE,
-            'Content-Range': `bytes ${OFFSET}-${endByte}/*`,
-          });
+          assert(reqOpts.headers);
+          assert.equal(reqOpts.headers['Content-Length'], CHUNK_SIZE);
+          assert.equal(
+            reqOpts.headers['Content-Range'],
+            `bytes ${OFFSET}-${endByte}/*`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+          );
 
           const data = await getAllDataFromRequest();
 
@@ -1132,10 +1168,18 @@ describe('resumable-upload', () => {
           await up.startUploading();
 
           const endByte = CONTENT_LENGTH - NUM_BYTES_WRITTEN + OFFSET - 1;
-          assert.deepEqual(reqOpts.headers, {
-            'Content-Length': CONTENT_LENGTH - NUM_BYTES_WRITTEN,
-            'Content-Range': `bytes ${OFFSET}-${endByte}/${CONTENT_LENGTH}`,
-          });
+          assert(reqOpts.headers);
+          assert.equal(
+            reqOpts.headers['Content-Length'],
+            CONTENT_LENGTH - NUM_BYTES_WRITTEN
+          );
+          assert.equal(
+            reqOpts.headers['Content-Range'],
+            `bytes ${OFFSET}-${endByte}/${CONTENT_LENGTH}`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+          );
           const data = await getAllDataFromRequest();
 
           assert.equal(data.byteLength, CONTENT_LENGTH - NUM_BYTES_WRITTEN);
@@ -1259,6 +1303,47 @@ describe('resumable-upload', () => {
 
       up.responseHandler(RESP);
     });
+
+    it('currentInvocationId.chunk should be different after success', done => {
+      const beforeCallInvocationId = up.currentInvocationId.chunk;
+      const RESP = {data: '', status: 200};
+      up.on('uploadFinished', () => {
+        assert.notEqual(beforeCallInvocationId, up.currentInvocationId.chunk);
+        done();
+      });
+      up.responseHandler(RESP);
+    });
+
+    it('currentInvocationId.chunk should be the same after error', done => {
+      const beforeCallInvocationId = up.currentInvocationId.chunk;
+      const RESP = {data: {error: new Error('Error.')}};
+      up.destroy = () => {
+        assert.equal(beforeCallInvocationId, up.currentInvocationId.chunk);
+        done();
+      };
+      up.responseHandler(RESP);
+    });
+  });
+
+  it('currentInvocationId.offset should be different after success', async () => {
+    const beforeCallInvocationId = up.currentInvocationId.offset;
+    up.makeRequest = () => {
+      return {};
+    };
+    await up.getAndSetOffset();
+    assert.notEqual(beforeCallInvocationId, up.currentInvocationId.offset);
+  });
+
+  it('currentInvocationId.offset should be the same on error', async done => {
+    const beforeCallInvocationId = up.currentInvocationId.offset;
+    up.destroy = () => {
+      assert.equal(beforeCallInvocationId, up.currentInvocationId.offset);
+      done();
+    };
+    up.makeRequest = () => {
+      throw new Error() as GaxiosError;
+    };
+    await up.getAndSetOffset();
   });
 
   describe('#getAndSetOffset', () => {
@@ -1271,10 +1356,12 @@ describe('resumable-upload', () => {
       up.makeRequest = async (reqOpts: GaxiosOptions) => {
         assert.strictEqual(reqOpts.method, 'PUT');
         assert.strictEqual(reqOpts.url, URI);
-        assert.deepStrictEqual(reqOpts.headers, {
-          'Content-Length': 0,
-          'Content-Range': 'bytes */*',
-        });
+        assert(reqOpts.headers);
+        assert.equal(reqOpts.headers['Content-Length'], 0);
+        assert.equal(reqOpts.headers['Content-Range'], 'bytes */*');
+        assert.ok(
+          X_GOOG_API_HEADER_REGEX.test(reqOpts.headers['x-goog-api-client'])
+        );
         done();
         return {};
       };
@@ -2089,9 +2176,16 @@ describe('resumable-upload', () => {
           assert(request.chunkWritesInRequest > 1);
 
           assert.equal(request.dataReceived, CONTENT_LENGTH);
-          assert.deepStrictEqual(request.opts.headers, {
-            'Content-Range': `bytes 0-*/${CONTENT_LENGTH}`,
-          });
+          assert(request.opts.headers);
+          assert.equal(
+            request.opts.headers['Content-Range'],
+            `bytes 0-*/${CONTENT_LENGTH}`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(
+              request.opts.headers['x-goog-api-client']
+            )
+          );
 
           done();
         });
@@ -2248,21 +2342,134 @@ describe('resumable-upload', () => {
               const endByte = offset + LAST_REQUEST_SIZE - 1;
 
               assert.equal(request.dataReceived, LAST_REQUEST_SIZE);
-              assert.deepStrictEqual(request.opts.headers, {
-                'Content-Length': LAST_REQUEST_SIZE,
-                'Content-Range': `bytes ${offset}-${endByte}/${CONTENT_LENGTH}`,
-              });
+              assert(request.opts.headers);
+              assert.equal(
+                request.opts.headers['Content-Length'],
+                LAST_REQUEST_SIZE
+              );
+              assert.equal(
+                request.opts.headers['Content-Range'],
+                `bytes ${offset}-${endByte}/${CONTENT_LENGTH}`
+              );
+              assert.ok(
+                X_GOOG_API_HEADER_REGEX.test(
+                  request.opts.headers['x-goog-api-client']
+                )
+              );
             } else {
               // The preceding chunks
               const endByte = offset + CHUNK_SIZE - 1;
 
               assert.equal(request.dataReceived, CHUNK_SIZE);
-              assert.deepStrictEqual(request.opts.headers, {
-                'Content-Length': CHUNK_SIZE,
-                'Content-Range': `bytes ${offset}-${endByte}/${CONTENT_LENGTH}`,
-              });
+              assert(request.opts.headers);
+              assert.equal(request.opts.headers['Content-Length'], CHUNK_SIZE);
+              assert.equal(
+                request.opts.headers['Content-Range'],
+                `bytes ${offset}-${endByte}/${CONTENT_LENGTH}`
+              );
+              assert.ok(
+                X_GOOG_API_HEADER_REGEX.test(
+                  request.opts.headers['x-goog-api-client']
+                )
+              );
             }
           }
+
+          done();
+        });
+
+        // init the request
+        upstreamBuffer.pipe(up);
+      });
+    });
+
+    describe('empty object', () => {
+      let uri = '';
+
+      beforeEach(() => {
+        uri = 'uri';
+
+        up.contentLength = 0;
+        up.createURI = (
+          callback: (error: Error | null, uri: string) => void
+        ) => {
+          up.uri = uri;
+          up.offset = 0;
+          callback(null, uri);
+        };
+      });
+
+      it('should support uploading empty objects', done => {
+        const CONTENT_LENGTH = 0;
+        const EXPECTED_NUM_REQUESTS = 1;
+
+        const upstreamBuffer = new Readable({
+          read() {
+            this.push(null);
+          },
+        });
+
+        const requests: {
+          dataReceived: number;
+          opts: GaxiosOptions;
+          chunkWritesInRequest: number;
+        }[] = [];
+        let overallDataReceived = 0;
+
+        up.makeRequestStream = async (opts: GaxiosOptions) => {
+          let dataReceived = 0;
+          let chunkWritesInRequest = 0;
+
+          const res = await new Promise(resolve => {
+            opts.body.on('data', (data: Buffer) => {
+              dataReceived += data.byteLength;
+              overallDataReceived += data.byteLength;
+              chunkWritesInRequest++;
+            });
+
+            opts.body.on('end', () => {
+              requests.push({dataReceived, opts, chunkWritesInRequest});
+
+              resolve({
+                status: 200,
+                data: {},
+              });
+
+              resolve(null);
+            });
+          });
+
+          return res;
+        };
+
+        up.on('error', done);
+
+        up.on('finish', () => {
+          // Ensure the correct number of requests and data look correct
+          assert.equal(requests.length, EXPECTED_NUM_REQUESTS);
+          assert.equal(overallDataReceived, CONTENT_LENGTH);
+
+          // Validate the single request
+          const request = requests[0];
+
+          assert.strictEqual(request.opts.method, 'PUT');
+          assert.strictEqual(request.opts.url, uri);
+
+          // We should be writing multiple chunks down the wire
+          assert(request.chunkWritesInRequest === 0);
+
+          assert.equal(request.dataReceived, CONTENT_LENGTH);
+          assert(request.opts.headers);
+
+          assert.equal(
+            request.opts.headers['Content-Range'],
+            `bytes 0-*/${CONTENT_LENGTH}`
+          );
+          assert.ok(
+            X_GOOG_API_HEADER_REGEX.test(
+              request.opts.headers['x-goog-api-client']
+            )
+          );
 
           done();
         });